requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
python-dotenv
asyncpg